Minor adjustments and improvements.
